EODHP-1004 airbus sar data adaptor integration

+++ resource_catalogue_fastapi/utils.py
@@ -103,27 +103,30 @@ def rate_limiter_dependency(workspace=Depends(get_workspace)):  # noqa: B008
         rate_limit(workspace)
 
 
-def upload_file_s3(body: str, bucket: str, key: str) -> bool:
+def upload_file_s3(body: str, bucket: str, key: str, error_on_exist: bool = False) -> bool:
     """Upload data to an S3 bucket. Returns a bool indicating whether the file existed previously"""
     s3_client = boto3.client("s3")
-    file_existed = False
+    file_exists = False
 
     try:
         # Check if the file already exists
         s3_client.head_object(Bucket=bucket, Key=key)
-        file_existed = True
+        file_exists = True
     except ClientError as e:
         # If a 404 error is raised, the file does not exist
         if e.response["Error"]["Code"] != "404":
             logging.error(f"Error checking if file exists: {e}")
             raise
+    if error_on_exist and file_exists:
+        logging.error(f"File already exists: {key}")
+        raise FileExistsError(f"File already exists: {key}")
     try:
         s3_client.put_object(Body=body, Bucket=bucket, Key=key)
     except ClientError as e:
         logging.error(f"File upload failed: {e}")
         raise
 
-    return file_existed
+    return file_exists
 
 
 def delete_file_s3(bucket: str, key: str):
@@ -196,6 +199,7 @@ def execute_order_workflow(
     authorization: str,
     stac_key: str,
     workspace_bucket: str,
+    workspaces_domain: str,
 ):
     """Executes a data adaptor workflow in the provider's workspace as the given user with auth"""
 
@@ -212,8 +216,12 @@ def execute_order_workflow(
             "workspace": user_workspace,
             "stac_key": stac_key,
             "workspace_bucket": workspace_bucket,
+            "workspace_domain": workspaces_domain,
+            "env": "prod",
         }
     }
+
+    logger.info(f"Sending request to {url} with payload: {payload}")
 
     response = requests.post(url, headers=headers, json=payload)
     response.raise_for_status()
